+++ cmd/hidden_lake/applications/messenger/pkg/config/config.go
@@ -10,7 +10,6 @@ func GetConfigSettings(pCfg config.IConfig) SConfigSettings {
 	return SConfigSettings{
 		SConfigSettings: config.SConfigSettings{
 			FMessagesCapacity: sett.GetMessagesCapacity(),
-			FWorkSizeBits:     sett.GetWorkSizeBits(),
 			FLanguage:         language.FromILanguage(sett.GetLanguage()),
 		},
 	}